requirements.txt
# Polymarket Volatility Scalping Dashboard
# =========================================

# API & HTTP
requests>=2.28.0

# Data Processing
pandas>=1.5.0
numpy>=1.23.0

# Dashboard (Dash + Plotly)
dash>=2.14.0
dash-bootstrap-components>=1.5.0
plotly>=5.18.0

# AI Analysis
openai>=1.6.0

# Environment & Config
python-dotenv>=1.0.0
pyyaml>=6.0.0

# Scheduling
schedule>=1.2.0
apscheduler>=3.10.0

# Database
# sqlite3 is built-in Python

# Utilities
python-dateutil>=2.8.0
